third_party/blink/web_tests/external/wpt/html/semantics/forms/the-label-element/label-attributes.sub-expected.txt

This is a testharness.js-based test.
[FAIL] A div element which contains labelable element is removed.
  assert_equals: The number of labels should be 1 after the labelable element is removed but label element is still in the same tree. expected 1 but got 0
[FAIL] A labelable element not in a document can label element in the same tree.
  assert_equals: The number of labels associated with a form control should be the number of label elements for which it is a labeled control. expected 2 but got 0
Harness: the test ran to completion.
